functional mod

+++ Utilities/Currency.cs
@@ -9,7 +9,7 @@ namespace SpiritMod.Utilities
 	{
 		public GlyphCurrency(int coinItemID, long currencyCap) : base(coinItemID, currencyCap) { }
 
-		public override void GetPriceText(string[] lines, ref int currentLine, int price)
+		public override void GetPriceText(string[] lines, ref int currentLine, long price)
 		{
 			Color glyphColor = Color.Orange * (Main.mouseTextColor / 255f);
 			lines[currentLine++] = string.Format("[c/{0:X2}{1:X2}{2:X2}:{3} {4} {5}]",
